Rewrote most of makeme (the pavement.py file).
The main reason was to get rid of the paver requirement, and just use
Python's built in argument parsing (which is actually quite nice).

Most of the code is the same, but manually rewritten to make sure we
only do things we want to do. Some old and unused imports have been
removed, same for unused functions.

It could still probably be cleaned up, but it's more specific to the
needs now, as opposed to before.